Bumped htmlunit and jenkins to 3.0 and 2.361.4
Refactored MoveActionTest (affected by namesapces change  htmlunit 2->3
from com.gargoylesoftware to org.htmlunit)
so each test have its own rule, so it no longer suffer random concurency
issue during clean up
