+++ pom.xml
@@ -4,7 +4,7 @@
     <parent>
         <groupId>org.jenkins-ci.plugins</groupId>
         <artifactId>plugin</artifactId>
-        <version>4.40</version> <!-- jenkins core 2.200 requires 4.0 or higher -->
+        <version>4.66</version> <!-- jenkins core 2.200 requires 4.0 or higher -->
         <relativePath />
     </parent>
     <groupId>io.jenkins.plugins</groupId>
@@ -12,7 +12,7 @@
     <version>1.4.4-SNAPSHOT</version>
     <packaging>hpi</packaging>
     <properties>
-        <jenkins.version>2.341</jenkins.version>
+        <jenkins.version>2.361.4</jenkins.version>
         <useBeta>true</useBeta> <!--can be deleted when manage permission does not require it anymore. https://github.com/jenkinsci/jenkins/blob/master/core/src/main/java/jenkins/model/Jenkins.java -->
     </properties>
     <name>Simple Queue Plugin</name>
