+++ src/test/java/cz/mendelu/xotradov/test/moves/package-info.java
@@ -1,0 +1,9 @@
+package cz.mendelu.xotradov.test.moves;
+/**
+ * There is an concurency error somewhere. When the tests were run as pasrt of single class,
+ * sometimes some leftover was not garbagecollected in time, and thus one or two tests randomly failed
+ * on cleanup, because the *rule* could not clean between runs.
+ *
+ * To ahve each of them in own file, with own RULE is much slower, but the concurency issue moreover disapeared
+ * Still, may appear from time to time:(
+ */
